PS1 icon cleanup (Part 1)
PS1 needs lots of cleanup.
They need to be sorted correctly and renamed. (Partially done in this commit)
The icons all need to be resized.
The collection needs to be verified as complete.

The size of the PS1 library is far too large for one person to do all of these jobs.